--- a/config/config_test.go
+++ b/config/config_test.go
@@ -30,7 +30,7 @@ func TestReadFromFile(t *testing.T) {
 func TestWriteToFile(t *testing.T) {
 	path := filepath.Join(os.TempDir(), "config.yaml")
 	t.Log(path)
-	cfg := DefaultConfig()
+	cfg := DefaultConfigForTesting()
 	if err := cfg.WriteToFile(path); err != nil {
 		t.Errorf("error writing config: %s", err.Error())
 		return
@@ -80,7 +80,7 @@ func TestWriteToFileWithExtraData(t *testing.T) {
 }
 
 func TestConfigSummaryString(t *testing.T) {
-	summary := DefaultConfig().SummaryString()
+	summary := DefaultConfigForTesting().SummaryString()
 	t.Log(summary)
 	if !strings.Contains(summary, "API") {
 		t.Errorf("expected summary to list API port")
@@ -88,7 +88,7 @@ func TestConfigSummaryString(t *testing.T) {
 }
 
 func TestConfigGet(t *testing.T) {
-	cfg := DefaultConfig()
+	cfg := DefaultConfigForTesting()
 	cases := []struct {
 		path   string
 		expect interface{}
@@ -130,7 +130,8 @@ func TestConfigSet(t *testing.T) {
 	}
 
 	for i, c := range cases {
-		cfg := DefaultConfig()
+		cfg := DefaultConfigForTesting()
+
 		err := cfg.Set(c.path, c.value)
 		if !(err == nil && c.err == "" || err != nil && err.Error() == c.err) {
 			t.Errorf("case %d error mismatch. expected: %s, got: %s", i, c.err, err)
@@ -155,7 +156,7 @@ func TestConfigSet(t *testing.T) {
 }
 
 func TestImmutablePaths(t *testing.T) {
-	dc := DefaultConfig()
+	dc := DefaultConfigForTesting()
 	for path := range ImmutablePaths() {
 		if _, err := dc.Get(path); err != nil {
 			t.Errorf("path %s default configuration error: %s", path, err.Error())
@@ -164,39 +165,39 @@ func TestImmutablePaths(t *testing.T) {
 }
 
 func TestConfigValidate(t *testing.T) {
-	if err := DefaultConfig().Validate(); err != nil {
+	if err := DefaultConfigForTesting().Validate(); err != nil {
 		t.Errorf("error validating config: %s", err)
 	}
 
 	//  cases that should fail:
-	p := DefaultConfig()
+	p := DefaultConfigForTesting()
 
 	// Profile:
 	p.Profile = nil
 	if err := p.Validate(); err == nil {
 		t.Error("When given no Profile, config.Validate did not catch the error.")
 	}
-	p.Profile = DefaultProfile()
+	p.Profile = DefaultProfileForTesting()
 	p.Profile.Type = "badType"
 	if err := p.Validate(); err == nil {
 		t.Error("When given bad input in Profile, config.Validate did not catch the error.")
 	}
 	// Repo:
-	r := DefaultConfig()
+	r := DefaultConfigForTesting()
 	r.Repo.Type = "badType"
 	if err := r.Validate(); err == nil {
 		t.Error("When given bad input in Repo, config.Validate did not catch the error.")
 	}
 
 	// Store:
-	s := DefaultConfig()
+	s := DefaultConfigForTesting()
 	s.Store.Type = "badType"
 	if err := s.Validate(); err == nil {
 		t.Error("When given bad input in Store, config.Validate did not catch the error.")
 	}
 
 	// Logging:
-	l := DefaultConfig()
+	l := DefaultConfigForTesting()
 	l.Logging.Levels["qriapi"] = "badType"
 	if err := l.Validate(); err == nil {
 		t.Error("When given bad input in Logging, config.Validate did not catch the error.")
@@ -207,7 +208,7 @@ func TestConfigCopy(t *testing.T) {
 	cases := []struct {
 		config *Config
 	}{
-		{DefaultConfig()},
+		{DefaultConfigForTesting()},
 	}
 	for i, c := range cases {
 		cpy := c.config.Copy()