This is a huge speedup! The tests in config/ go from about 40 seconds to 2.
    
Before this change:

```
Run 0
ok      github.com/qri-io/qri/api       11.612s
ok      github.com/qri-io/qri/cmd       62.384s
ok      github.com/qri-io/qri/config    37.921s
ok      github.com/qri-io/qri/lib       13.243s
ok      github.com/qri-io/qri/p2p        4.397s
Run 1
ok      github.com/qri-io/qri/api       10.292s
ok      github.com/qri-io/qri/cmd       47.731s
ok      github.com/qri-io/qri/config    38.941s
ok      github.com/qri-io/qri/lib       15.966s
ok      github.com/qri-io/qri/p2p        4.017s
```

After this change:

```
Run 0
ok      github.com/qri-io/qri/api       15.766s
ok      github.com/qri-io/qri/cmd       54.987s
ok      github.com/qri-io/qri/config     2.260s
ok      github.com/qri-io/qri/lib       12.002s
ok      github.com/qri-io/qri/p2p        2.390s
Run 1
ok      github.com/qri-io/qri/api       10.640s
ok      github.com/qri-io/qri/cmd       57.889s
ok      github.com/qri-io/qri/config     1.274s
ok      github.com/qri-io/qri/lib       13.730s
ok      github.com/qri-io/qri/p2p        2.762s
```

Depends upon #530.